fix(DraggableCore): fix passing style
When using `React.cloneElement(this.props.children, props), we used to
override style by spreading it into another object. When that was removed,
we simply passed style on, unmodified. It appears some quirk of `cloneElement`
causes the render cycle to actually lose the value of `this.props.children.props.style`,
causing it to pass `undefined`.

This, of course, breaks just about everything.

Because this is affecting dependent packages I am pushing an update now then will
add a test to catch this.